Moved MessageBus to utils
