+++ src/main/java/com/gamesense/client/command/commands/AutoReplyCommand.java
@@ -1,7 +1,7 @@
 package com.gamesense.client.command.commands;
 
 import com.gamesense.client.command.Command;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.module.modules.misc.AutoReply;
 
 public class AutoReplyCommand extends Command{
