+++ src/main/java/com/gamesense/client/command/commands/FixGuiCommand.java
@@ -3,7 +3,7 @@ package com.gamesense.client.command.commands;
 import com.gamesense.client.GameSenseMod;
 import com.gamesense.client.clickgui.GameSenseGUI;
 import com.gamesense.client.command.Command;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 
 /**
  * @Author Hoosiers on 10/03/20
