+++ src/main/java/com/gamesense/client/command/commands/OpenFolderCommand.java
@@ -1,7 +1,7 @@
 package com.gamesense.client.command.commands;
 
 import com.gamesense.client.command.Command;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 
 import java.awt.*;
 import java.io.File;
