+++ src/main/java/com/gamesense/client/command/commands/PrefixCommand.java
@@ -1,7 +1,7 @@
 package com.gamesense.client.command.commands;
 
 import com.gamesense.client.command.Command;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 
 public class PrefixCommand extends Command{
