+++ src/main/java/com/gamesense/client/command/commands/SetSettingCommand.java
@@ -3,7 +3,7 @@ package com.gamesense.client.command.commands;
 import com.gamesense.api.settings.Setting;
 import com.gamesense.client.GameSenseMod;
 import com.gamesense.client.command.Command;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.module.Module;
 import com.gamesense.client.module.ModuleManager;
