+++ src/main/java/com/gamesense/client/command/commands/ToggleCommand.java
@@ -1,6 +1,6 @@
 package com.gamesense.client.command.commands;
 
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 import com.mojang.realmsclient.gui.ChatFormatting;
 import com.gamesense.client.command.Command;
 import com.gamesense.client.module.ModuleManager;
