+++ src/main/java/com/gamesense/client/commands2/CommandManager.java
@@ -1,5 +1,6 @@
 package com.gamesense.client.commands2;
 
+import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.commands2.commands.*;
 
 import java.util.ArrayList;
