+++ src/main/java/com/gamesense/client/module/modules/combat/AutoWeb.java
@@ -8,7 +8,7 @@ import com.gamesense.api.util.players.friends.Friends;
 import com.gamesense.api.settings.Setting;
 import com.gamesense.api.util.world.BlockUtils;
 import com.gamesense.api.util.world.EntityUtil;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.module.Module;
 
 import com.gamesense.client.module.ModuleManager;
