+++ src/main/java/com/gamesense/client/module/modules/exploits/CoordExploit.java
@@ -2,7 +2,7 @@ package com.gamesense.client.module.modules.exploits;
 
 import com.gamesense.api.settings.Setting;
 import com.gamesense.api.util.world.WorldUtils;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.module.Module;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.math.Vec3d;
