+++ src/main/java/com/gamesense/client/module/modules/misc/DiscordRPCModule.java
@@ -1,7 +1,7 @@
 package com.gamesense.client.module.modules.misc;
 
 import com.gamesense.api.util.misc.Discord;
-import com.gamesense.client.commands2.MessageBus;
+import com.gamesense.api.util.misc.MessageBus;
 import com.gamesense.client.module.Module;
 
 public class DiscordRPCModule extends Module {
